only add rtorrent.rc if not already present

+++ rtorrent-compile.sh
@@ -81,9 +81,11 @@ if [ -f "$USER_HOME/start_rtorrent.php" ]; then
 	sudo su $SUDO_USER -c "php $USER_HOME/start_rtorrent.php"
 fi
 
-# Install .rtorrent.rc
-wget --no-check-certificate https://raw.githubusercontent.com/jonnyboy/rtorrent-scripts/master/config/rtorrent.rc -O $USER_HOME/.rtorrent.rc
-sed -i -e "s/USERNAME/$SUDO_USER/" $USER_HOME/.rtorrent.rc
+# Install .rtorrent.rc if not already exists
+if [ -f "$USER_HOME/.rtorrent.rc" ]; then
+	wget --no-check-certificate https://raw.githubusercontent.com/jonnyboy/rtorrent-scripts/master/config/rtorrent.rc -O $USER_HOME/.rtorrent.rc
+	sed -i -e "s/USERNAME/$SUDO_USER/" $USER_HOME/.rtorrent.rc
+fi
 mkdir -p $USER_HOME/Downloads/.session/
 mkdir -p $USER_HOME/Downloads/watch
 mkdir -p $USER_HOME/Downloads/incomplete
